Make results processor idempotent
After loading the report into memory and before any uploading, the
processor first asks /api/runs if a TestRun corresponding to the report
has already been created and returns early if yes. This is safe to do
because 1) test run creation is the very last step in processor and 2)
raw_results_url is now a unique identifier after #54.

The processor now returns two different success codes: 201 if a report
is uploaded and a test run is created; 200 if the test run already
exists and hence nothing is done. This allows easier filtering of logs.

Fixes #247.